create integrate project
